LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/button-layout/flex-expected.txt


1
2
3
4
 
1
2
3
4

1
2
3
4
 
abc

abc

PASS inline-flex 
PASS flex 
FAIL align-items:flex-start should work assert_equals: expected 84 but got 74
FAIL align-items:stretch should work assert_equals: expected 13 but got 33
